requirements.txt
discord
python-dotenv
aiohttp
